[claude] relocate cross-module AST helpers to ast_helpers.py (#133)
