+++ .uncoded/stubs/src/uncoded/stubs.pyi
@@ -4,7 +4,7 @@ import ast
 from collections.abc import Iterable
 from dataclasses import dataclass, field
 from pathlib import Path
-from uncoded.extract import property_kind
+from uncoded.ast_helpers import property_kind
 from uncoded.sync import remove_file, sync_file
 
 VALUE_WIDTH_CAP = 80
