+++ src/uncoded/extract.py
@@ -6,6 +6,8 @@ from collections.abc import Iterable, Iterator
 from dataclasses import dataclass, field
 from pathlib import Path
 
+from uncoded.ast_helpers import assign_target_name, property_kind
+
 
 @dataclass
 class ClassInfo:
@@ -26,33 +28,6 @@ class ModuleInfo:
     functions: list[str] = field(default_factory=list)
 
 
-def property_kind(
-    node: ast.FunctionDef | ast.AsyncFunctionDef,
-) -> str | None:
-    """Classify a method by its property-related decorators.
-
-    Returns "property" for @property, "setter" for @<name>.setter,
-    "deleter" for @<name>.deleter, or None for a plain method.
-    """
-    for d in node.decorator_list:
-        if isinstance(d, ast.Name) and d.id == "property":
-            return "property"
-        if isinstance(d, ast.Attribute) and d.attr in ("setter", "deleter"):
-            return d.attr
-    return None
-
-
-def _assign_target_name(node: ast.Assign | ast.AnnAssign) -> str | None:
-    """Return the single-name target of an assignment, or None if not a simple name."""
-    if isinstance(node, ast.AnnAssign):
-        target = node.target
-        return target.id if isinstance(target, ast.Name) else None
-    if len(node.targets) != 1:
-        return None
-    target = node.targets[0]
-    return target.id if isinstance(target, ast.Name) else None
-
-
 def extract_module(source: str, rel_path: str) -> ModuleInfo:
     """Parse Python source and extract classes, functions, and constants."""
     tree = ast.parse(source)
@@ -67,7 +42,7 @@ def extract_module(source: str, rel_path: str) -> ModuleInfo:
             methods: list[str] = []
             for n in ast.iter_child_nodes(node):
                 if isinstance(n, (ast.AnnAssign, ast.Assign)):
-                    name = _assign_target_name(n)
+                    name = assign_target_name(n)
                     if name:
                         attributes.append(name)
                 elif isinstance(n, (ast.FunctionDef, ast.AsyncFunctionDef)):
@@ -84,7 +59,7 @@ def extract_module(source: str, rel_path: str) -> ModuleInfo:
         elif isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef)):
             functions.append(node.name)
         elif isinstance(node, (ast.Assign, ast.AnnAssign)):
-            name = _assign_target_name(node)
+            name = assign_target_name(node)
             if name:
                 constants.append(name)
         elif isinstance(node, ast.TypeAlias):
